tests: split learn tests into focused files and extend round-trip coverage
